Move all shared methods from RK4SimulationStepper and RK6SimulationStepper into AbstractRKSimulationStepper. This only leaves the step() method.
